backend/requirements.txt
# backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
httpx
sentence-transformers
faiss-cpu
cachetools
beautifulsoup4
requests
pandas
lxml # pandas.read_html requires this
apscheduler
thefuzz